move parts of token to metadata project
