+++ Sustainsys.Saml2.Metadata/KeyType.cs
@@ -1,0 +1,9 @@
+﻿namespace Sustainsys.Saml2.Metadata
+{
+	public enum KeyType
+	{
+		Unspecified = 0,
+		Signing = 1,
+		Encryption = 2
+	}
+}
