+++ Sustainsys.Saml2.Metadata/Services/PassiveRequestorEndpoint.cs
@@ -1,0 +1,6 @@
+﻿namespace Sustainsys.Saml2.Metadata.Services
+{
+	public class PassiveRequestorEndpoint : Endpoint
+	{
+	}
+}
