+++ Sustainsys.Saml2.Metadata/Services/SingleSignOutNotificationEndpoint.cs
@@ -1,0 +1,6 @@
+﻿namespace Sustainsys.Saml2.Metadata.Services
+{
+	public class SingleSignOutNotificationEndpoint : Endpoint
+	{
+	}
+}
