+++ Sustainsys.Saml2.Metadata/Sustainsys.Saml2.Metadata.csproj
@@ -23,8 +23,4 @@
     <PackageReference Include="System.Security.Cryptography.Xml" Version="4.5.0" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Descriptors\" />
-  </ItemGroup>
-
 </Project>
