+++ Sustainsys.Saml2.Metadata/Tokens/SecurityAlgorithms.cs
@@ -1,4 +1,4 @@
-﻿namespace Sustainsys.Saml2.Tokens
+﻿namespace Sustainsys.Saml2.Metadata.Tokens
 {
 	public static class SecurityAlgorithms
 	{
