+++ Sustainsys.Saml2.Metadata/Tokens/SecurityKey.cs
@@ -1,4 +1,4 @@
-﻿namespace Sustainsys.Saml2.Tokens
+﻿namespace Sustainsys.Saml2.Metadata.Tokens
 {
 	public abstract class SecurityKey
 	{
