+++ Sustainsys.Saml2.Metadata/Tokens/X509IssuerSerialKeyIdentifierClause.cs
@@ -1,10 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Security.Cryptography.X509Certificates;
 using System.Numerics;
+using System.Security.Cryptography.X509Certificates;
 
-namespace Sustainsys.Saml2.Tokens
+namespace Sustainsys.Saml2.Metadata.Tokens
 {
     public class X509IssuerSerialKeyIdentifierClause : SecurityKeyIdentifierClause
     {
