--- Sustainsys.Saml2/Metadata/KeyType.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Sustainsys.Saml2.Metadata
-{
-	public enum KeyType
-	{
-		Unspecified = 0,
-		Signing = 1,
-		Encryption = 2
-	}
-}
