+++ Sustainsys.Saml2/Metadata/MetadataSerializer.cs
@@ -11,8 +11,10 @@ using System.Numerics;
 using System.Security.Cryptography;
 using System.Security.Cryptography.X509Certificates;
 using System.Xml;
+using Sustainsys.Saml2.Metadata.Descriptors;
 using Sustainsys.Saml2.Metadata.Localization;
 using Sustainsys.Saml2.Metadata.Services;
+using Sustainsys.Saml2.Metadata.Tokens;
 using SigningCredentials = Microsoft.IdentityModel.Tokens.SigningCredentials;
 
 namespace Sustainsys.Saml2.Metadata
@@ -73,8 +75,8 @@ namespace Sustainsys.Saml2.Metadata
 		protected virtual AffiliationDescriptor CreateAffiliationDescriptorInstance() =>
 			new AffiliationDescriptor();
 
-		protected virtual ApplicationService CreateApplicationServiceEndpointInstance() =>
-			new ApplicationService();
+		protected virtual ApplicationServiceEndpoint CreateApplicationServiceEndpointInstance()
+            => new ApplicationServiceEndpoint();
 
 		protected virtual ApplicationServiceDescriptor CreateApplicationServiceInstance() =>
 			new ApplicationServiceDescriptor();
@@ -2871,7 +2873,7 @@ namespace Sustainsys.Saml2.Metadata
 		protected virtual SingleSignOutNotificationEndpoint ReadSingleSignOutNotificationEndpoint(XmlReader reader) =>
 			ReadWrappedEndpoint(reader, CreateSingleSignOutNotificationEndpointInstance);
 
-		protected virtual ApplicationService ReadApplicationServiceEndpoint(XmlReader reader) =>
+		protected virtual ApplicationServiceEndpoint ReadApplicationServiceEndpoint(XmlReader reader) =>
 			ReadWrappedEndpoint(reader, CreateApplicationServiceEndpointInstance);
 
 		protected virtual PassiveRequestorEndpoint ReadPassiveRequestorEndpoint(XmlReader reader) =>
