--- Sustainsys.Saml2/Metadata/PassiveRequestorEndpoint.cs
@@ -1,8 +1,0 @@
-﻿using Sustainsys.Saml2.Metadata.Services;
-
-namespace Sustainsys.Saml2.Metadata
-{
-	public class PassiveRequestorEndpoint : Endpoint
-	{
-	}
-}
