--- Sustainsys.Saml2/Metadata/SingleSignOutNotificationEndpoint.cs
@@ -1,8 +1,0 @@
-﻿using Sustainsys.Saml2.Metadata.Services;
-
-namespace Sustainsys.Saml2.Metadata
-{
-	public class SingleSignOutNotificationEndpoint : Endpoint
-	{
-	}
-}
