+++ Sustainsys.Saml2/Selectors/SecurityTokenResolver.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Text;
 using System.Xml;
 using Sustainsys.Saml2.Configuration;
+using Sustainsys.Saml2.Metadata.Tokens;
 using Sustainsys.Saml2.Tokens;
 
 namespace Sustainsys.Saml2.Selectors
